requirements.txt
boto
django-debug-toolbar
django-extensions
fabric
ipython
mock
python-dateutil==1.5
south
yolk

-e git+http://github.com/django/django.git#egg=django
-e svn+https://ctctwspylib.svn.sourceforge.net/svnroot/ctctwspylib#egg=ctctwspylib
